refactor dependency scanning for reuse when building lockfile
