+++ src/luarocks/dir.lua
@@ -6,6 +6,7 @@ local core = require("luarocks.core.dir")
 
 dir.path = core.path
 dir.split_url = core.split_url
+dir.normalize = core.normalize
 
 --- Strip the path off a path+filename.
 -- @param pathname string: A path+name, such as "/a/b/c"
@@ -26,18 +27,6 @@ end
 function dir.dir_name(pathname)
    assert(type(pathname) == "string")
    return (pathname:gsub("/*$", ""):match("(.*)[/]+[^/]*")) or ""
-end
-
---- Normalize a url or local path.
--- URLs should be in the "protocol://path" format. System independent
--- forward slashes are used, removing trailing and double slashes
--- @param url string: an URL or a local pathname.
--- @return string: Normalized result.
-function dir.normalize(name)
-   local protocol, pathname = dir.split_url(name)
-   pathname = pathname:gsub("\\", "/"):gsub("(.)/*$", "%1"):gsub("//", "/")
-   if protocol ~= "file" then pathname = protocol .."://"..pathname end
-   return pathname
 end
 
 --- Returns true if protocol does not require additional tools.
